Merge scope block for for init and the inner block
This stops you from declaring a variable inside a for
loop with the same name as one declared in the init
expression.

All statements that contain scoped statements (even
without an inner scope, such as if (true) int x = 3;)
now all use the ScopeBlock param that combines the
declarations for the outer and inner statements and
the vec of statements, so they're consistent with
for loops.

As if / while made their own scope before this, the only
change for them is not making two scopes.